Add flag --enable-go-utils to configure
This flag is off by default.  Controls whether golang binaries in directory
go-utils will be built. If the flag is on, configure will also require that
the go toolchain is installed.

This connects the top-level Makefile to the go-utils directory.

Previous behavior was python-centric, so there was no need to compile in gpMgmt,
so there was no 'all' target in the gpMgmt/Makefile.

Current behavior is to compile golang during the 'all' target in gpMgmt, and copy
the binary during the 'install' target.

Signed-off-by: Larry Hamel <lhamel@pivotal.io>
Signed-off-by: Tushar Dadlani <tdadlani@pivotal.io>